+++ core/src/main/java/nl/nn/adapterframework/pipes/TimeoutGuardPipe.java
@@ -141,20 +141,6 @@ public class TimeoutGuardPipe extends FixedForwardPipe {
 		return input.toString();
 	}
 
-	private String getParameterValue(ParameterValueList pvl,
-			String parameterName) {
-		ParameterList parameterList = getParameterList();
-		if (pvl != null && parameterList != null) {
-			for (int i = 0; i < parameterList.size(); i++) {
-				Parameter parameter = parameterList.getParameter(i);
-				if (parameter.getName().equalsIgnoreCase(parameterName)) {
-					return pvl.getParameterValue(i).asStringValue(null);
-				}
-			}
-		}
-		return null;
-	}
-
 	public void setThrowException(boolean b) {
 		throwException = b;
 	}